删除栏目下还有子栏目的情况，编辑时，将子栏目选为父栏目的情况

+++ 07/bool/admin/catedel.php
@@ -26,11 +26,26 @@ require('../include/init.php');
 
 $cat_id = $_GET['cat_id'] + 0;
 
+
+/*
+判断该栏目是否有子栏目  如果有，则不允许删除
+*/
+
+
+
 $cat = new CatModel();
+
+
+$sons = $cat->getSon($cat_id);
+print_r($sons);
+if (!empty($sons)) {
+  exit('有子栏目，不允许删除');
+}
+
+
 
 if($cat->delete($cat_id)) {
     echo '删除成功';
 } else {
     echo '删除失败';
 }
-
